refactor: Split webpack.config.base
