--- config/webpack.config.disable-contexts.js
@@ -1,9 +1,0 @@
-'use strict'
-
-const webpack = require('webpack')
-
-module.exports = {
-  plugins: [
-    new webpack.IgnorePlugin(/^\.\.?\/contexts/)
-  ]
-}
